Store the final URL reached by PhantomJS
The result is saved to port.final_url.  Transmitting it from PhantomJS is a
little hacky; the URL is logged first, and then the image, so we can split on
"\n" to separate the two.  Currently this splitting happens in
ScreenshotResults, but it could be done just as easily in ScreenshotWorker.