Add initial Nessus support
This code is potentially still a bit flakey. There are [deliberately]
no links to it in other parts of the web interface yet, and you’ll have
to modify app/workers/nessus_worker.rb to contain your Nessus creds.
This may be of some use to someone, or I will clean it up more in the
(near?) future.