Add GPLv3 This has been intended for a while, and the answer to the licensing question whenever it had come up, but I never got around to putting the actual text in the repository.
The reason for using the GPLv3 is to prevent people from selling a
hosted copy of Nepenthes without giving access to any improvements
they've made to the people who are paying for the hosted copy. If
you can make a strong argument for alternative licensing, contact
me before the list of contributors gets much larger. The GPLv3 has
been OK'd by all contributors in discussions with them.