Fix SSL checking
Pass an empty input to the OpenSSL command, and check the exit status.

The previous approach of looking for `SSL-Session` is no longer
reliable, since that gets printed in some cases even if we didn't
establish a TLS connection. Instead, pass an empty string as the input
so a successful connection will terminate immediately. This allows us to
determine if the port uses TLS by looking at the return value:

0:   Successful connection
1:   Error of some sort
124: Timed out

Also, if we found a certificate in the output, mark the port as using
TLS. This can be important if the port fails to meet OpenSSL's security
standards or requires a client certificate, in which case the return
code will indicate an error. It's possible that we should use this as
our sole indicator; I'm not sure if there are any cases where a server
certificate will not be displayed for a valid connection.

Remaining issues:

- If a TLS port fails to respond quickly enough, it will be classified
  as non-TLS. I think this is unavoidable because many non-TLS ports
  time out without causing errors.